pyproject.toml
[tool.poetry]
name = "otto-gateway-message-producer"
version = "1.0.0"
description = "A micoservice for Otto that sends all discord gateway event's to RabbitMQ"
authors = ["JacobMonck <jacobamonck@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
corded = "^1.5.0"
aio-pika = "^6.8.0"
python-dotenv = "^0.19.0"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
